+++ cricbox/match_statistics/models.py
@@ -21,12 +21,12 @@ class MatchStatistics(models.Model):
     match = models.OneToOneField(Match, on_delete=models.CASCADE, null=True)
     result = models.ForeignKey(Result, verbose_name="Result", on_delete=models.PROTECT, null=True)
     report = models.TextField("Report", blank=True, null=True)
-    london_fields_overs = models.PositiveSmallIntegerField("London Fields Overs", null=True)
+    london_fields_overs = models.DecimalField("London Fields Overs", max_digits=3, decimal_places=1, null=True)
     london_fields_score = models.PositiveSmallIntegerField("London Fields Score", null=True)
     london_fields_wickets = models.PositiveSmallIntegerField("London Fields Wickets", null=True)
     opposition_score = models.PositiveSmallIntegerField("Opposition score", null=True)
     opposition_wickets = models.PositiveSmallIntegerField("Opposition Wickets", null=True)
-    opposition_overs = models.PositiveSmallIntegerField("Opposition overs", null=True)
+    opposition_overs = models.DecimalField("Opposition overs", max_digits=3, decimal_places=1, null=True)
     london_fields_first_event = models.PositiveSmallIntegerField(choices=EVENT_FIRST, default=0, null=True)
     report_byline = models.CharField(max_length=200, null=True)
     report_author = models.CharField(max_length=200, null=True)
@@ -37,6 +37,7 @@ class MatchStatistics(models.Model):
 
     def get_batsman(self):
         return self.batsman_set.get_queryset()
+
 
     class Meta:
         ordering = ("match",)
